Refactor block/edit to use less dependencies
